13.12

+++ controllers/blogController.js
@@ -1,7 +1,13 @@
-const { Blog } = require("../models");
+const { Blog, User } = require("../models");
 
 exports.getAllBlogs = async (req, res, next) => {
-  const blogs = await Blog.findAll();
+  const blogs = await Blog.findAll({
+    attributes: { exclude: ["UserId"] },
+    include: {
+      model: User,
+      attributes: ["name"],
+    },
+  });
   res.json(blogs);
 };
 
@@ -30,7 +36,19 @@ exports.updateLikes = async (req, res, next) => {
 };
 
 exports.createBlog = async (req, res, next) => {
-  const blog = await Blog.create(req.body);
+  console.log("CURRENT USER");
+  console.log(req.user.dataValues);
+
+  const blog = new Blog({
+    author: req.body.author,
+    title: req.body.title,
+    url: req.body.url,
+    likes: req.body.likes,
+    UserId: req.user.dataValues.id,
+  });
+
+  await blog.save();
+
   res.status(201).json(blog);
 };
 
@@ -43,11 +61,19 @@ exports.deleteBlog = async (req, res) => {
       message: "No document with that ID",
     });
   }
-  await Blog.destroy({
-    where: {
-      id: blog.dataValues.id,
-    },
-  });
 
-  res.status(200).end();
+  if (req.user.dataValues.id === blog.UserId) {
+    await Blog.destroy({
+      where: {
+        id: blog.dataValues.id,
+      },
+    });
+
+    res.status(200).end();
+  } else {
+    res.status(401).json({
+      status: "fail",
+      message: "You can delete only your own blogs",
+    });
+  }
 };
